#276 move ipc to its own mobule
